+++ crates/cubecl-cuda/src/compute/io/mod.rs
@@ -1,4 +1,1 @@
 pub(crate) mod controller;
-
-mod base;
-pub use base::*;
